SOC-798 | Helios authorization - move code to separate module
